python-behave/requirements.txt
# BDD Framework
behave==1.2.6
behave-html-formatter==0.9.10

# Browser Automation
playwright==1.49.1

# Configuration
python-dotenv==1.0.1

# Assertions
assertpy==1.1

# Utilities
Faker==33.1.0

# Reporting
allure-behave==2.13.5
